requirements.txt
streamlit
pandas
numpy
matplotlib
mysql-connector-python
bcrypt